비밀번호 인증 후 평가 삭제 및 수정 기능 추가

+++ BiBlet/src/main/java/a/b/c/controller/AppraisalController.java
@@ -68,34 +68,45 @@ public class AppraisalController {
 	 * 도서 상세보기
 	 */
 	@PostMapping("/read")
-	public String writeComment(int actionFlag, 
+	public String writeComment(int actionFlag, Model model, 
 			@ModelAttribute("insertCmd") InsertCmd insertCmd,
 			@ModelAttribute("deleteCmd") DeleteCmd deleteCmd, 
-			@ModelAttribute("updateCmd") UpdateCmd updateCmd, Model model) throws UnsupportedEncodingException {
-		
-		System.out.println("actionFlag:" + actionFlag);
-		AppraisalVO appraisal = new AppraisalVO();
-		BookShelfVO bookShelf = new BookShelfVO();
+			@ModelAttribute("updateCmd") UpdateCmd updateCmd) throws UnsupportedEncodingException {
 		
 		// 테스트 하기 전마다 회원 등록 후 평가작성을 하지 않은 새로운 회원번호로 진행해야함
 		MemberVO member = new MemberVO();
-		Long mem_num = (long) 16; // 테스트용 회원 번호(현재 테이블에 6번회원까지 있음)
+		Long mem_num = (long) 15; // 테스트용 회원 번호(현재 테이블에 6번회원까지 있음)
 		member.setMem_num(mem_num);
 
 		String redirectUrl = "";
 		if(actionFlag == 1) {
 			String encodedParam = URLEncoder.encode(insertCmd.getQuery(), "UTF-8");
 			redirectUrl = writeComment(insertCmd, mem_num) + encodedParam;
+			return redirectUrl;
+			
 		}else if(actionFlag == 2) {
 			String encodedParam = URLEncoder.encode(deleteCmd.getQuery(), "UTF-8");
-			redirectUrl = deleteComment(deleteCmd, mem_num) + encodedParam;
+			
+			if(deleteCmd.getMem_pass().equals(deleteCmd.getPassCheck())) {
+				System.out.println("삭제 전 비밀번호 확인 "+ deleteCmd.getMem_pass().equals(deleteCmd.getPassCheck()));
+				redirectUrl = deleteComment(deleteCmd, mem_num) + encodedParam;
+				return redirectUrl;
+			}
+			
+			
 		}else if(actionFlag == 3) {
 			String encodedParam = URLEncoder.encode(updateCmd.getQuery(), "UTF-8");
+			
+			if(updateCmd.getMem_pass().equals(updateCmd.getPassCheck())) {
+				System.out.println("수정 전 비밀번호 확인 "+ updateCmd.getMem_pass().equals(updateCmd.getPassCheck()));
+				redirectUrl = updateComment(updateCmd, mem_num) + encodedParam;
+				return redirectUrl;
+			}
+			
 			redirectUrl = updateComment(updateCmd, mem_num) + encodedParam;
+			return redirectUrl;
 		}
-
 		return redirectUrl;
-
 	}
 	
 	/**
